api/requirements.txt
fastapi
uvicorn
pydantic
pillow
numpy
torch